refactor(project): reorganize project layout

+++ engine/src/textfile.cpp
@@ -29,7 +29,7 @@ char* textFileRead(char *fn)
             if (count > 0)
             {
                 content = (char *)malloc(sizeof(char) * (count + 1));
-                count = fread(content, sizeof(char), count, file);
+                count = static_cast<int>(fread(content, sizeof(char), count, file));
                 content[count] = '\0';
             }
             fclose(file);
